chore: cleanup code noice
